+++ model/fields_test.go
@@ -1,0 +1,12 @@
+package model
+
+import "testing"
+
+func TestFields(t *testing.T) {
+	// TODO: IDField.GetID, SetID
+	// TODO: CodeField.GetCode
+	// TODO: CreatedAtField.GetDateTime
+	// TODO: IsActiveField.GetIsActive, SetIsActive
+	// TODO: Item.Initialize
+	// TODO: AutoItem.Initialize
+}