+++ model/model.go
@@ -1,0 +1,2 @@
+// Package model contains Schema type and methods
+package model